chore: rename core/Parser.ts file & followup #10460
