+++ packages/core/src/index.ts
@@ -22,7 +22,7 @@ export { Hub, HubCapsule, HubCallback, HubPayload } from './Hub';
 export { I18n } from './I18n';
 export * from './JS';
 export { Signer } from './Signer';
-export * from './Parser';
+export * from './parseAWSExports';
 export * from './Providers';
 export { FacebookOAuth, GoogleOAuth } from './OAuthHelper';
 export * from './RNComponents';
