+++ packages/core/src/parseAWSExports.ts
